requirements.txt

Django>=4.1.4,<5.0
djangorestframework==3.15.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
python-dotenv==1.0.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
pytz==2023.4
